move slug field control to @wordpress/fields package
