+++ packages/edit-site/src/style.scss
@@ -1,4 +1,5 @@
 @import "../../dataviews/src/style.scss";
+@import "../../fields/src/styles.scss";
 
 @import "./components/add-new-template/style.scss";
 @import "./components/block-editor/style.scss";
@@ -29,7 +30,6 @@
 @import "./components/style-book/style.scss";
 @import "./components/editor-canvas-container/style.scss";
 @import "./components/post-edit/style.scss";
-@import "./components/post-fields/style.scss";
 @import "./components/post-list/style.scss";
 @import "./components/resizable-frame/style.scss";
 @import "./hooks/push-changes-to-global-styles/style.scss";
