+++ packages/fields/README.md
@@ -14,7 +14,13 @@ npm install @wordpress/fields --save
 
 <!-- START TOKEN(Autogenerated API docs) -->
 
-Nothing to document.
+### SlugEdit
+
+Undocumented declaration.
+
+### SlugView
+
+Undocumented declaration.
 
 <!-- END TOKEN(Autogenerated API docs) -->
