+++ packages/fields/src/fields/index.ts
@@ -1,0 +1,2 @@
+export { default as SlugEdit } from './slug/slug.edit';
+export { default as SlugView } from './slug/slug.view';
