+++ packages/fields/src/fields/slug/style.scss
