+++ packages/fields/src/fields/slug/types.ts
@@ -1,0 +1,5 @@
+export type SlugItem = {
+	permalink_template?: string;
+	link?: string;
+	slug?: string;
+};
