+++ packages/fields/src/styles.scss
@@ -1,0 +1,1 @@
+@import "./fields/slug/style.scss";
