+++ packages/fields/tsconfig.json
@@ -6,6 +6,19 @@
 		"declarationDir": "build-types",
 		"checkJs": false
 	},
-	"references": [],
+	"references": [
+		{ "path": "../components" },
+		{ "path": "../compose" },
+		{ "path": "../data" },
+		{ "path": "../element" },
+		{ "path": "../i18n" },
+		{ "path": "../icons" },
+		{ "path": "../primitives" },
+		{ "path": "../private-apis" },
+		{ "path": "../warning" },
+		{ "path": "../url" },
+		{ "path": "../notices" },
+		{ "path": "../dataviews" }
+	],
 	"include": [ "src" ]
 }
